Move thrift things together, and make thrift serializer for pickle
